Refactoring, modularization, removed JSON reader, added sample test
